requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
dnspython==2.2.0
idna==3.3
numpy==1.22.2
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
